fix(modeline): fix the performance degradation with active region in long buffer
In long buffers we called `line-number-at-pos` twice to determine the
length of the selected range.  Unfortunately this function is extremly
slow and degrades rapidly in performance for long buffers.

It was replaced by `format-mode-line` [1] which is implemented in C.
This results in near-instant performance.  Unfortunatelly, it only
works for visible buffers, but for our use-case of a mode-line
construct it is sufficient (and properly hacky in the Emacs
tradition).

[1] https://emacs.stackexchange.com/a/3822/23535